Fixed subsection links.
Subsection links are those that link to a specific part of another document. My subclass of Pod::Simple::XHTML makes them `rel="subection"`, so I updated the code that makes such links load within the tree to match any rel attribute that ends with "section". I also added a bit of code to capture the location hash so that it properly scrolls to it. Yay!