Update for latest commits to Pod::Simple.
Allison accepted a variation on my patch for striping prefixes from verbatim
blocks. It's not automatic as I originall proposed, but must be supplied as a
code reference to calculate the amount to strip for each verbatim block. I had
to add the `batch_mode_page_object_init()` method to my Pod::Site::XHTML
subclass in order to properly configure this setting, because
Pod::Simple::HTMLBatch calls this method in order for the formatting class to
do any configuration it needs. I've also future-proofed it, since
Pod::Simple::XHTML doesn't implement this method, but might at a future date,
at which time we'll need to properly call it.

I've also upped the requirement to Pod::Simple 3.09, since that will be the
first version with the new feature.