Add tests for Pod::Site::XHTML output.
Fix a bug where the URL for local links was borked to the full file name,
which is not what we want. Change it to the proper URI including the base_uri
(using the first of those passed), as that looks much better, is indexable,
and the jQuery-powered stuff uses the name attribute request the document
anyway.

The use of the base URI requires a bit of a hack with a `local`ed variable,
but this is a perfect use for it, even if it's kind of ugly.

Also remove an ancient JavaScript function call in `body onload` that hasn't
been used in a long time.